+++ ui/src/components/agents/AgentsSidebar.tsx
@@ -27,11 +27,6 @@ import type { SessionUser } from "@/lib/api"
 import type { AgentSource, AgentThread } from "@/lib/agents/types"
 import type { SidebarLayout } from "@/components/sidebar-layout"
 import { SidebarUserMenu } from "@/components/SidebarUserMenu"
-import {
-  ReviewSidebarPanel,
-  ReviewSidebarProvider,
-  useReviewSidebarData,
-} from "@/components/agents/ReviewSidebar"
 import { Button } from "@/components/ui/button"
 import {
   SidebarCollapseButton,
@@ -114,7 +109,6 @@ export function AgentsSidebar({
   useSeedAgentThreadDetails(visibleThreads, activeThreadId)
   useRunCompletionNotifier(visibleThreads, activeThreadId)
   const groups = groupThreads(activeThreads)
-  const reviewSidebar = useReviewSidebarData()
 
   return (
     <SidebarFrame
@@ -164,44 +158,40 @@ export function AgentsSidebar({
         })}
       </nav>
 
-      {reviewSidebar ? (
-        <ReviewSidebarPanel data={reviewSidebar} />
-      ) : (
-        <div className="min-h-0 flex-1 overflow-y-auto px-2 pb-2">
-          <ThreadGroup
-            label="Today"
-            threads={groups.today}
-            activeThreadId={activeThreadId}
-            onNavigate={layout.closeOnMobile}
-          />
-          <ThreadGroup
-            label="Last 7 days"
-            threads={groups.last7}
-            activeThreadId={activeThreadId}
-            onNavigate={layout.closeOnMobile}
-            defaultCollapsed
-          />
-          <ThreadGroup
-            label="Last 30 days"
-            threads={groups.last30}
-            activeThreadId={activeThreadId}
-            onNavigate={layout.closeOnMobile}
-            defaultCollapsed
-          />
-          <ThreadGroup
-            label="Older"
-            threads={groups.older}
-            activeThreadId={activeThreadId}
-            onNavigate={layout.closeOnMobile}
-          />
-          <ResolvedThreadGroup
-            threads={resolvedThreads}
-            hasMore={resolvedHasMore}
-            activeThreadId={activeThreadId}
-            onNavigate={layout.closeOnMobile}
-          />
-        </div>
-      )}
+      <div className="min-h-0 flex-1 overflow-y-auto px-2 pb-2">
+        <ThreadGroup
+          label="Today"
+          threads={groups.today}
+          activeThreadId={activeThreadId}
+          onNavigate={layout.closeOnMobile}
+        />
+        <ThreadGroup
+          label="Last 7 days"
+          threads={groups.last7}
+          activeThreadId={activeThreadId}
+          onNavigate={layout.closeOnMobile}
+          defaultCollapsed
+        />
+        <ThreadGroup
+          label="Last 30 days"
+          threads={groups.last30}
+          activeThreadId={activeThreadId}
+          onNavigate={layout.closeOnMobile}
+          defaultCollapsed
+        />
+        <ThreadGroup
+          label="Older"
+          threads={groups.older}
+          activeThreadId={activeThreadId}
+          onNavigate={layout.closeOnMobile}
+        />
+        <ResolvedThreadGroup
+          threads={resolvedThreads}
+          hasMore={resolvedHasMore}
+          activeThreadId={activeThreadId}
+          onNavigate={layout.closeOnMobile}
+        />
+      </div>
 
       <div className="p-2">
         <SidebarUserMenu user={user} showSettingsLink />
@@ -546,17 +536,15 @@ export function AgentsShell({
 }) {
   const layout = useSidebarLayout()
   return (
-    <ReviewSidebarProvider>
-      <SidebarLayoutProvider value={layout}>
-        <div className="agents-ui flex h-svh overflow-hidden bg-[var(--ui-bg)]">
-          <AgentsSidebar
-            user={user}
-            activeThreadId={activeThreadId}
-            layout={layout}
-          />
-          <div className="flex min-w-0 flex-1">{children}</div>
-        </div>
-      </SidebarLayoutProvider>
-    </ReviewSidebarProvider>
+    <SidebarLayoutProvider value={layout}>
+      <div className="agents-ui flex h-svh overflow-hidden bg-[var(--ui-bg)]">
+        <AgentsSidebar
+          user={user}
+          activeThreadId={activeThreadId}
+          layout={layout}
+        />
+        <div className="flex min-w-0 flex-1">{children}</div>
+      </div>
+    </SidebarLayoutProvider>
   )
 }
